Split components (#1)
* update: introduce pages & components

 fix lint

* update: starting split components
